backend/requirements.txt
django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
django-extensions==3.1.5
